Move session card to its own file
